Move canonical path resolution to a package
