+++ packages/canonical-path/src/content/path-resolvers.js
