+++ packages/canonical-path/src/index.js
@@ -1,0 +1,7 @@
+const content = require('./content');
+const section = require('./section');
+
+module.exports = {
+  content,
+  section,
+};
