+++ services/graphql-server/package.json
@@ -8,6 +8,7 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
+    "@base-cms/canonical-path": "^0.6.0",
     "@base-cms/db": "^0.6.0",
     "@base-cms/embedded-media": "^0.6.0",
     "@base-cms/html": "^0.6.0",
