+++ services/graphql-server/src/graphql/resolvers/website/section.js
@@ -1,12 +1,8 @@
 const { BaseDB } = require('@base-cms/db');
-const { isFunction: isFn, cleanPath } = require('@base-cms/utils');
+const { section: canonicalPathFor } = require('@base-cms/canonical-path');
 const getProjection = require('../../utils/get-projection');
 const getGraphType = require('../../utils/get-graph-type');
 const { createTitle, createDescription } = require('../../utils/website-section');
-
-const pathResolvers = {
-  alias: section => section.alias,
-};
 
 const loadHierarchy = async (section, load, projection, sections = []) => {
   const ref = BaseDB.get(section, 'parent');
@@ -23,21 +19,7 @@ module.exports = {
    *
    */
   WebsiteSection: {
-    canonicalPath: async (section, _, ctx) => {
-      const { canonicalRules } = ctx;
-      const { websiteSection: sectionRules } = canonicalRules;
-      const { parts, prefix } = sectionRules;
-
-      const values = await Promise.all(parts.map((key) => {
-        const fn = pathResolvers[key];
-        return isFn(fn) ? fn(section, ctx) : section[key];
-      }));
-
-      const path = cleanPath(values.filter(v => v).map(v => String(v).trim()).join('/'));
-      if (!path || path === 'home') return '/';
-      if (prefix) return `/${cleanPath(prefix)}/${path}`;
-      return `/${path}`;
-    },
+    canonicalPath: (section, _, ctx) => canonicalPathFor(section, ctx),
 
     /**
      * Placeholder.
